LIBASPACE-360. Replaced "_result" partial with "_result_records_summary"
In ArchivesSpace v3.4.0, the "_result.html.erb" file was modified
to use the "_result_records_summary.html.erb" partial.

To preserve the UMD "Found in" customization, copied the
"_result_records_summary.html.erb" partial for ArchivesSpace 3.5.1 into
this plugin, and modified to move the "Found in" field to be the first
field.

This had a side effect (due to the partial's use of the
"_result_breadcrumbs.html.erb" partial, that a resource icon is now
shown next to the resource name.

After these changes, the UMD-customized "_results.html.erb" partial
was no longer needed, so it was removed (restoring the use if the
stock ArchivesSpace "_results.html.erb" partial.

Updated the "PublicInterfaceCustomization.md" documentation.

https://issues.umd.edu/browse/LIBASPACE-360